+++ steam_api/Steamworks/Implementation/SteamAppList.cs
@@ -9,25 +9,25 @@ namespace SKYNET.Steamworks.Implementation
     [StructLayout(LayoutKind.Sequential)]
     public class SteamAppList : ISteamInterface
     {
-        public int GetAppBuildId(IntPtr nAppID)
+        public int GetAppBuildId(uint nAppID)
         {
             Write("GetAppBuildId");
             return 10;
         }
 
-        public int GetAppInstallDir(IntPtr nAppID, string pchDirectory, int cchNameMax)
+        public int GetAppInstallDir(uint nAppID, string pchDirectory, int cchNameMax)
         {
             Write("GetAppInstallDir");
             return -1;
         }
 
-        public int GetAppName(IntPtr nAppID, string pchName, int cchNameMax)
+        public int GetAppName(uint nAppID, string pchName, int cchNameMax)
         {
             Write("GetAppName\n");
             return -1;
         }
 
-        public uint GetInstalledApps(IntPtr pvecAppID, uint unMaxAppIDs)
+        public uint GetInstalledApps(uint pvecAppID, uint unMaxAppIDs)
         {
             Write("GetInstalledApps\n");
             return 0;
